Move generate_entity_id to entity helpers
