+++ homeassistant/helpers/__init__.py
@@ -3,22 +3,7 @@ Helper methods for components within Home Assistant.
 """
 import re
 
-from homeassistant.const import (
-    ATTR_ENTITY_ID, CONF_PLATFORM, DEVICE_DEFAULT_NAME)
-from homeassistant.util import ensure_unique_string, slugify
-
-
-def generate_entity_id(entity_id_format, name, current_ids=None, hass=None):
-    """ Generate a unique entity ID based on given entity IDs or used ids. """
-    name = name.lower() or DEVICE_DEFAULT_NAME.lower()
-    if current_ids is None:
-        if hass is None:
-            raise RuntimeError("Missing required parameter currentids or hass")
-
-        current_ids = hass.states.entity_ids()
-
-    return ensure_unique_string(
-        entity_id_format.format(slugify(name.lower())), current_ids)
+from homeassistant.const import CONF_PLATFORM
 
 
 def validate_config(config, items, logger):
